Move condition module
