+++ gui/model/parameter/condition.py
